requirements.txt
aiofiles==23.2.1
altair==5.3.0
annotated-types==0.6.0
anyio==3.7.1
attrs==23.2.0
black==24.4.2
blinker==1.8.2
cachetools==5.3.3
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
distro==1.9.0
dnspython==2.6.1
docx2txt==0.8
email_validator==2.1.1
faiss-cpu==1.8.0
fastapi==0.111.0
fastapi-cli==0.0.2
filelock==3.14.0
fsspec==2024.3.1
gitdb==4.0.11
GitPython==3.1.43
h11==0.14.0
hiredis==2.3.2
httpcore==1.0.5
httptools==0.6.1
httpx==0.27.0
huggingface-hub==0.23.0
idna==3.7
igraph==0.11.5
Jinja2==3.1.3
joblib==1.4.2
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
llvmlite==0.42.0
lxml==5.2.1
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
mpmath==1.3.0
mypy-extensions==1.0.0
mysqlclient==2.2.4
networkx==3.3
nltk==3.8.1
numba==0.59.1
numpy==1.26.3
openai==1.3.3
orjson==3.10.3
packaging==24.0
pandas==2.2.2
pathspec==0.12.1
pillow==10.3.0
platformdirs==4.2.1
plotly==5.22.0
protobuf==4.25.3
pyarrow==16.0.0
pydantic==2.7.1
pydantic_core==2.18.2
pydeck==0.9.0
Pygments==2.18.0
pynndescent==0.5.12
pypdf==4.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-magic==0.4.27
python-multipart==0.0.9
python-pptx==0.6.23
pytz==2024.1
PyYAML==6.0.1
redis==5.0.4
referencing==0.35.1
regex==2024.4.28
requests==2.31.0
rich==13.7.1
rpds-py==0.18.1
safetensors==0.4.3
scikit-learn==1.4.2
scipy==1.13.0
sentence-transformers==2.7.0
sentencepiece==0.2.0
shellingham==1.5.4
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
SQLAlchemy==2.0.29
starlette==0.37.2
streamlit==1.34.0
sympy==1.12
tenacity==8.2.3
texttable==1.7.0
threadpoolctl==3.5.0
tiktoken==0.5.1
tokenizers==0.15.2
toml==0.10.2
toolz==0.12.1
torch==2.3.0
torchvision==0.18.0
tornado==6.4
tqdm==4.66.4
transformers==4.38.1
typer==0.12.3
typing_extensions==4.11.0
tzdata==2024.1
ujson==5.9.0
umap-learn==0.5.5
urllib3==1.26.6
uvicorn==0.29.0
uvloop==0.19.0
watchfiles==0.21.0
websockets==12.0
XlsxWriter==3.2.0
